# app/docs/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    FileType:
      type: string
      enum:
      - LIBRE_DOCUMENT
      - LIBRE_PRESENTATION
      - LIBRE_SPREADSHEET
      - MS_DOCUMENT
      - MS_PRESENTATION
      - MS_SPREADSHEET
    InsertFile:
      type: object
      properties:
        filename:
          type: string
        destinationFolderId:
          type: string
        type:
          $ref: "#/components/schemas/FileType"
      required:
      - destinationFolderId
      - type
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
paths:
  /services/docs/files/create:
    post:
      parameters:
      - name: Cookie
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertFile"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Create File
      tags:
      - Files Resource
  /services/docs/files/open/{nodeId}:
    get:
      parameters:
      - name: nodeId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: offset_from_utc
        in: query
        schema:
          type: integer
          format: int32
      - name: redirect
        in: query
        schema:
          type: boolean
      - name: version
        in: query
        schema:
          type: integer
          format: int32
      - name: Cookie
        in: header
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Open File
      tags:
      - Files Resource
  /services/docs/wopi/{nodeId}:
    get:
      parameters:
      - name: nodeId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: access_token
        in: query
        schema:
          type: string
      - name: offset_from_utc
        in: query
        schema:
          type: integer
          format: int32
      - name: version
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Docs Editor Attributes
      tags:
      - Wopi Resource
  /services/docs/wopi/{nodeId}/contents:
    post:
      parameters:
      - name: nodeId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: access_token
        in: query
        schema:
          type: string
      - name: offset_from_utc
        in: query
        schema:
          type: integer
          format: int32
      - name: Content-Length
        in: header
        schema:
          type: integer
          format: int64
      - name: X-COOL-WOPI-IsAutosave
        in: header
        schema:
          type: boolean
      - name: X-COOL-WOPI-IsExitSave
        in: header
        schema:
          type: boolean
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Save Blob
      tags:
      - Wopi Resource
    get:
      parameters:
      - name: nodeId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: access_token
        in: query
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/octet-stream: {}
      summary: Get Blob
      tags:
      - Wopi Resource
info:
  title: carbonio-docs-connector-ce-app API
  version: unversioned
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value